walkdetector: log "walk_root" as a &str
Summary:
This makes it easier to capture by the tracing logger.

The logger implements tracing::field::Visit to visit the tracing event's fields, but I didn't implement Visit::record_debug() for now because I didn't want to encourage unstructured/untyped data. I had imagine the `%` convenience tracing operator presents the field as a str, but that isn't the case, so just convert explicitly to a str.

Reviewed By: quark-zju

Differential Revision: D75874174

fbshipit-source-id: 509b94e936ff86dc94dd8e7ae724d7312035424d