set heurystic to creation token and set creation token to 1
Summary:
We only support full repo bundles now. If we want to support full repo bundles for `git init && git fetch` we should set the creation token so we don't consider bundles for further fetches, apart from the initial one.

This test is absolutely necessary to know that sometimes creationToken has to be camelCase and sometimes it absolute cannot.

Reviewed By: RajivTS

Differential Revision: D75790678

fbshipit-source-id: 9527d854d5e7087f1d26349512a41e3a2156ee0e