Run tailer only if in master region
Summary: This diff adds logic that makes it so only if the tailer is collocated with master it will run the tailer. This is checked during shard assingment. In the next diff I will make it so we check it periodically, in case leadership changes regions while the process is running

Reviewed By: RajivTS

Differential Revision: D75869779

fbshipit-source-id: a500338b612890457f17f59f0a6ae33230736893