Fix symlink traversal and also keep track of symlinks traversed
Summary:
There was a bug in the clap syntax for the follow-symlinks option that
was preventing it from working correctly.  This diff makes it so that
the traversal code defaults to not following symlinks unless the command line
(--follow-symlinks) is given.

Also added feature for tracking the number of symlinks traversed if follow
symlinks is set.

Reviewed By: giorgidze

Differential Revision: D75892503

fbshipit-source-id: 2474b034b49a364cf0e7ac13481b84a5297d6c0b