Traversal debug Thrift feature now works with repos starting with fbsource
Summary:
The --read-file-via feature was failing on my fbsource directory since my repo
is named 'fbsource-eden' and not just 'fbsource'. Modified the matching code to work with
repos that start with 'fbsource' instead of the more strict 'fbsource'.

Might be interesting to know what other variations are used and whether we can be even more
generic.

Reviewed By: giorgidze

Differential Revision: D75762650

fbshipit-source-id: 340e57d79abcaab92ba81f4d5fd6de9604de54f6