Integration test demonstrating support for multiple tags with same name
Summary:
The earlier versions of `gitimport` didn't used to peel tags when trying to upload the tag contents. With that version, it was possible to have the below scenario work for gitimport but fail for Mononoke Git server:
- Create tag A point to commit C
- Created another tag with the same name A and point it to the object id of the old tag A
- gitimport the repo

The earlier version of gitimport would only import the new tag A without uploading the git object corresponding to the old tag A. On Mononoke server when performing the clone, we would try to peel the tags and return everything referenced by it only to find the inner tag missing and would error out. This actually happened in production for `whatsapp/android` repo since it was imported early last year.

In the current version of gitimport, this bug is fixed so this integration test serves as protection against regression and as a proof that we support this particular use-case

Reviewed By: markbt

Differential Revision: D75964395

fbshipit-source-id: 29835270ea9915b4cccf938fdc4881ade5fc607b