Add upload tag functionality to gitimport binary
Summary: Every now and then we come across a weird tag object that gitimport as part of its normal execution doesn't upload for some reason. This new command is created specifically for that purpose. Note that this won't add a mapping in `bonsai_tag_mapping` table but will simply upload the tag object to the storage layer. We might want to extend this to add the mappings as well in the future, if we see a need.

Reviewed By: markbt

Differential Revision: D75960452

fbshipit-source-id: 81596968cd5d5c4861c533c0a7608411a8426485