histedit: abort when attempting to fold subtree commits
Summary:
Currently, `histedit` does not combine the subtree metadata when it folds subtree commits,
so the subtree metadata can be lost. I’m not sure that simply merging the metadata would
fully solve the issue. Until we have a proper fix, this diff aborts `histedit` whenever a user
tries to fold subtree commits.

Reviewed By: muirdm

Differential Revision: D75897810

fbshipit-source-id: 3a29885d97a414afe646142ef82dd36299a4204a