github: fix base image path in Dockerfile
Summary:
I accidentally used the x86_64 image. It should be changed to the
multi-platform image so arm64 can build just fine.

Note this uses qemu, which seems to slow down the arm build by 4x (2min ->
8min). However, it generates a "multi-platform" image, which is convenient.

If we'd like to use native arm, it is easy to use "matrix" to build 2 images
per platform. However, it's tricky (but still possible [1]) to build the
multi-platform image.

[1]: https://docs.docker.com/build/ci/github-actions/multi-platform/

Reviewed By: muirdm

Differential Revision: D75983166

fbshipit-source-id: b71e762bff958e76bd68c830b440ccffb8ff2eb1