Add timeout to gh merge queue query
Summary:
When checking for the merge queue support, we've seen cases of the command hanging. Let's just add a timeout to kill the spawned proccess. This makes sure we keep this reasonable.

I went with 10 seconds. It's a very simple query, so if it takes more than 1-2 seconds something is wrong.

Worst case scenario: the timeout hits when it shouldn't and the repo is detected as not supporting merge queue, which is a minor feature. It would still work otherwise.

Reviewed By: muirdm

Differential Revision: D76067308

fbshipit-source-id: ab14eecc86629f72273f58095661a0ece57c4c3e