walkdetector: always record interesting directory metadata
Summary:
Previously if we saw a humongous directory that was already covered by a higher up directory walk, we would not retain the directory size hints. However, the directory size is still important as we see deeper activity under the large directory to avoid marking the large directory as "walked" too aggressively.

Tweak things so we always record interesting directory metadata (even with the interesting directory is already covered by a walk).

Reviewed By: quark-zju

Differential Revision: D76012258

fbshipit-source-id: 7783fb7c11a169e865639498a43639341ed63623