types: add optimized RepoPath::depth() method
Summary:
Use SIMD enabled bytecount crate to provide optimized depth() method. This counts the number of path components, and is equivalent to `path.components().count()`.

On mac this gave ~8x speedup from 52ns to 7ns for a medium length path. On linux it gave ~20x speedup from 62ns to 3ns.

The walkdetector library checks the depth of a path a lot, and the nanoseconds add up. This optimization sped up a benchmark of the walk detector against a big list of files by 20% (7s -> 5.5s).

Reviewed By: quark-zju

Differential Revision: D76063718

fbshipit-source-id: eddd99b90c2ed358184a8289521248455bdb3aec