walkdetector: stop passing around "now" value
Summary:
Passing "now" everywhere was a side effect of the previous lame way of mocking Instant::now. Now with mock_instant, we can just use the global, convenient Instant::now().

In particular I want to add more fine-grained GC to avoid needing "full" collections very often, and not having to pass "now" everywhere makes it a lot more convenient when checking if things are expired.

There is some tiny increased overhead of accessing the system clock more, but we are talking about nanoseconds (I hope?), so the least of our worries.

Reviewed By: quark-zju

Differential Revision: D76012262

fbshipit-source-id: 735feed9eb316501ca585dd4496ba0adaa5f56f6