walkdetector: raise default GC interval
Summary: Raise default GC interval from 1 second to 10 seconds. After recent improvements, we don't require a "full" GC for correct expiration behavior, so we can run it less frequently. The GC can take a significant amount of milliseconds when there is a lot of walk activity going on (regardless of whether it collects anything), so it is better to not run it frequently.

Reviewed By: quark-zju

Differential Revision: D76012259

fbshipit-source-id: e1553c1a8ec40f58be63b554d7026e9c853782b7