walkdetector: don't retain metadata for empty directories
Summary:
If we see "interesting" directory size hints, we retain the hints since they will likely be useful when making walk detection decisions in the future for that directory. Previously we counted "empty" as interesting, but it isn't actually very interesting since empty directories will never see walk activity. It is advantageous to avoid storing boring hints since it avoids keeping "boring" nodes in our data structure.

Note that the directories are not completely empty (since we don't track empty directories), but more likely have zero directories and a normal amount of files, or vice versa.

Reviewed By: quark-zju

Differential Revision: D76012251

fbshipit-source-id: ff24c88645791a09861e01c827fd6ba4363ccf4a