walkdetector: don't merge into deep grandparent walk
Summary: I noticed under certain circumstances two sibling/cousin walks can merge into their grandparent walk when the grandparent walk already contains them. The intention of the heuristic is to advance the grandparent walk to contain the cousins - it wasn't expecting for the grandparent walk to already contain the cousins. This can have the effect of "jumping" past the higher threshold of a very large directory inadvertently (see test for example).

Reviewed By: quark-zju

Differential Revision: D76012248

fbshipit-source-id: 08ee3785f831e878df5243a78f20912c84730d82