walkdetector: support splitting off child walks
Summary:
Normally walk detection wants to coalesce walks together to consolidate things as quickly as possible. But when there is a big imbalance in the tree being walked, it can be beneficial to split off child walks, knowing that things will coalesce better into the child walk.

For example (simplification):

          root
          /   \
     small    big
     /   \    / \
    a     b  c   d
            / \   \
           e  f    g
                  / \
                 h   i

As things are walked, we quickly coalesce to a walk at "root" of depth=1 (including "a", "b", "c", and "d"). But as we see further activity under "big", we won't advance "root" anymore because it is waiting for activity under multiple children (i.e. under "small" and "big"). However, we can split off a separate walk for just "big", and then all the activity under big can propagate up to the walk rooted at "big".

We do the splitting by tracking how many times "small" and "big" have depth advancing walks under them (previously we only tracked whether "small" and "big" had any advancing walks). When we see that "big" has enough advancements, we split off a new walk for "big".

Another approach to solving this case is to notice that "small" has been 100% searched, so can be ignored when considering if we should advance "root". This however requires relatively "perfect" knowledge about the total number of files/directories, which we don't necessarily have. Splitting off the child works for other cases such as "root" having lots of deep directories, but we only end up walking deep in "big" (i.e. "root" is walked shallowly a few levels, and then "big" is walked deeply).

Reviewed By: quark-zju

Differential Revision: D76012247

fbshipit-source-id: d7647ad7e6c4a57479d1c46a75fb387a71d7a364