Take an argument for the max number of files to read
Summary:
fbsource/fbcode is 8M files. Let's add an option so that when we run
on fbcode (or any other large dir), we can limit the number of files read.

PROOF OF CONCEPT - it's buggy and due to the recursion of traverse_directory,
it reads more files than strictly necessary (but only by a bit).

Reviewed By: giorgidze

Differential Revision: D75973166

fbshipit-source-id: 14e2d62b7b1328f788a017a00a5a8966c83fc39f