Short term, tactical bug fix to the --best-effort flag
Summary:
# Context

The `--best-effort` flag of the `sl cloud sync` command is used when cloud sync is triggered in the background. The flag documentation says that in this mode cloud sync should make the best effort to not to acquire the locks. Using this flag for background invocations has the correct and good intention.

As the test plan below demonstrates this is currently not the case. The background invocations even with the `--best-effort` flag keep running forever until they can acquire the lock. We can see in `dev_command_timers` table that this happens very often on the user's machines with stuck cloud sync processes in the background.

# This diff

Provides a short-term tactical fix. Avoids waiting for acquiring the lock if the lock is already held, avoids background processes getting stuck.

Note, the longer term we should make the implementation lock-free, opportunistically tagging lmvasquezg

Reviewed By: quark-zju

Differential Revision: D76101117

fbshipit-source-id: 709400c2dbc87758cf9ffe96ab3aaeb078263d23