derivation_queue: streamify derivation queue summary
Summary:
Currently `DerivationQueue::summary` queries information about all items in the queue before returning a result. This is usually fine with the queue size usually being in the double digits. However, for cases where the queue has grown unexpectedly large, this makes the summary command useless for debugging.

Streamify the return type of `DerivationQueue::summary` to allow quickly viewing a small number of the items in the queue in such cases.

Reviewed By: liubov-dmitrieva

Differential Revision: D76194173

fbshipit-source-id: ca1513475173c80695537f6a487c029e192f33cf