move EdenFS TW clone utility into its own contbuild and new subfolder
Summary:
We got complaints from a team depending on this package that our contbuild was stuck and stable was on a bpd-ed fbpkg. I noticed that the conveyor contbuild for this was running the whole eden test integration suite, which isn't needed. This diff is an attempt to isolate this fbpkg to reduce our support load - nothing changes in this file and were moving teams off of it so this is to just get this contbuild in a state where we don't need to respond to these types of requests.

AFAIK we don't use the eden contbuild file anymore (we use configerator based stuff now) so removing the `fbpkg_builder` part from this file should be safe.

In addition, I moved relevant files to their own subdir so i can more tightly scope the contbuild tests ran. This fbpkg doesn't need eden integration tests ran for example.

https://www.internalfb.com/conveyor/scm_client_infra/packages/releases
https://www.internalfb.com/fbpkg/package/fb.eden.tw.clone

Reviewed By: evanwex

Differential Revision: D75921440

fbshipit-source-id: d8be563fdc45f0b603fb78feefb680ebe26e8847