native Rust version of edenfs_config_manager -- part 6, implement fetching configs directly from configerator
Summary:
This diff adds a new fetch_* method for fetching the config from configerator.

This will most likely be the method used to fetch the config by local (ie. linux) sandcastle workers.

This is not plumbed through to the CLI.

So far, I've been following the design of the Python edenfs_config_manager.  I think it's time to discuss changes to make it more amenable for use as a library.  In particular, instead of a single EdenfsConfigManager::run() method that fetches, filters, and writes the config (and takes the parameters for all three), split those out into separate structs / functions.

Reviewed By: genevievehelsel

Differential Revision: D75537446

fbshipit-source-id: 5549b53c89099533225ec5567ee51752c34c1d6d