native Rust version of edenfs_config_manager -- part 8, write config to file
Summary:
This diff implements the code to atomically write the config to a file (or to stdout, if --dry-run is specified).

As I mentioned earlier in this stack, I think we probably should split fetching, filtering, and writing the config out of ::run(), to make it easier to use the code as a library.  But saving that for later.

Reviewed By: genevievehelsel

Differential Revision: D75574190

fbshipit-source-id: 8aa887212c378552e609af0dbaef731ff457d687