use last used time instead of atime in NFS GC
Summary:
## Context
`atime` is not accurate on NFS. It only updates for read file and read directory NFS commands. Also, the parents of an inode didn't get updated through `updateAtime()`. NFS GC needs a more accurate access time to depend on it when it wants to drop the not recently used inodes. Otherwise, we could get lots of ESTALE error due to dropping recently used inodes. More details about the `atime` accuracy and replace it with something more accurate in this [post](https://fb.workplace.com/groups/edenfs/permalink/2749942928509255/) and the task T101484211

To have a more reliable GC on NFS, we need to define a more accurate way to find the last time that an inode get used.

## This diff
Replaces `atime` with more accurate `nfsLastUsedTime_`. Therefore, GC more accurately drop inodes which are not used recently

Reviewed By: genevievehelsel

Differential Revision: D76085675

fbshipit-source-id: 09127d8613e296513f3e86726c58653c3d488dfd