Add NfsLastUsedTime to inodes
Summary:
## Context
`atime` is not accurate on NFS. It only updates for read file and read directory NFS commands. Also, the parents of an inode didn't get updated through `updateAtime()`. NFS GC needs a more accurate access time to depend on it when it wants to drop the not recently used inodes. Otherwise, we could get lots of ESTALE error due to dropping recently used inodes. More details about the `atime` accuracy and replace it with something more accurate in this [post](https://fb.workplace.com/groups/edenfs/permalink/2749942928509255/) and the task T101484211

To have a more reliable GC on NFS, we need to define a more accurate way to find the last time that an inode get used.

## This diff
Adds `nfsLastUsedTime_` field to `InodeBase` This field only get updated on NFS commands. Also, through `propagateNfsLastUsedTimeToParents` we can guarantee that the last used time get updated all the way to the root for all parents.

Reviewed By: genevievehelsel

Differential Revision: D76001603

fbshipit-source-id: 97009e20e390136b900ba6398b5f74c75fdd6082