Add commit_hg_mutation_history implementation to get predecessors of a commit
Summary: We add the commit_hg_mutation_history implementation to fetch all the predecessors of a commit. Currently the params parameter is unused, this can be used in future to return different formats of responses. Currently we are only returning the predecessors upto the limit returned by the backend API, in future we have to go to the last of such predecessors and get their predecessors until there are no more. For ease of review this will be done in a different diff in the stack.

Reviewed By: markbt

Differential Revision: D75103356

fbshipit-source-id: 2b2df7b68f70c19d3ed6b84477f13116d73bc804