Integration test highlighting lack of support for tags with incorrect naming scheme
Summary:
For repos like `osmeta/external/swift-toolchain-sqlite`, even after importing them in Mononoke we can see that they never get in sync with its Metagit version. This is because there are certain tags with non-standard naming scheme that are returned as annotated tags by Metagit whereas Mononoke returns them as simple branches.

This is because Mononoke assumes that bookmarks that do not start with the `refs/tags` prefix or contain this prefix within some namespace, are not tags. Technically, that should be correct behaviour but like all things Git you can hand-weave such non-standard tags and the server is supposed to work well with them. This integration test exists to demonstrate that currently it doesn't.

Differential Revision: D76269920

fbshipit-source-id: 701cfdca641582808a651552b202d209c718a5e8