Add dtype field into checkout conflicts
Summary:
# Bug Description
The eden journal is used to record file change events.
When doing a checkout in force mode, files that are not committed may change if they conflict with files in the new tree. Currently, these files are recorded as an unclean path, but the type of change is not recorded.

Repro:
https://www.internalfb.com/tasks/?t=225707535

# This Diff

Add dtype to CheckoutConflict. This is used later to record the conflict into the journal.

# Technical Details

# Discussion Points

Reviewed By: ViniGupta08

Differential Revision: D75817758

fbshipit-source-id: bf1cfd1b7489b42d4557a8f4b859ba45a995c20d