Derive copy for ReadFileMethod
Summary: The `Copy` trait is derived to allow the `ReadFileMethod` enum to implement the `Copy` trait, which provides a cheap and efficient way to clone the enum values. This is particularly useful when working with enums that contain small amounts of data, as it allows for implicit cloning of the enum values without incurring a significant performance overhead.

Differential Revision: D76218551

fbshipit-source-id: 280f14a8e778d2a9831f4423b18eef5a344e8f15