Set the capacity of file_paths to max_files to avoid resizing and copying of large vectors
Summary:
This diff aims to reduce the resizing and copying of large vectors. To achieve this, we've made the following changes:

### Added `max_files` parameter to `InProgressTraversal` struct

We've added a new field `max_files` to the `InProgressTraversal` struct to set the capacity of the `file_paths` vector. This allows us to avoid resizing the vector as we add files, reducing unnecessary memory allocations and copies.

### Updated `new` method of `InProgressTraversal` struct

The `new` method of `InProgressTraversal` now takes `max_files` and `follow_symlinks` as additional parameters. This ensures that the `file_paths` vector is initialized with the correct capacity and that we can follow symbolic links during traversal.

### Updated `DEFAULT_MAX_NUMBER_OF_FILES_FOR_TRAVERSAL` constant

We've introduced a new constant `DEFAULT_MAX_NUMBER_OF_FILES_FOR_TRAVERSAL` with a value of 1024 * 1024. This is a safer default of avoiding scanning entire FBSource. Can be set to larger value when needed.

Differential Revision: D76216129

fbshipit-source-id: 6a90a5ba6d2e7d431d9c5c9d9d5e4bf51a52dc18