Fix path reference for opening diff view
Summary:
The server-side changes for opening the diff view are throwing an exception because they are unable to correctly import the path modules needed to determine the full file path.
{F1979031277}
Reason:
The 'path' module is a Node.js import, but this code is running in the browser context, not the Node.js context. Therefore, it should not import the path module at all. I thought we had a polyfill (which provides the same functionality in the browser context to make it work), but unfortunately, we do not.
Options:
Concatenate the paths as strings like before, but be careful about using forward slashes (/) versus backslashes ().

Reviewed By: evangrayk

Differential Revision: D76362167

fbshipit-source-id: caa4b3c3e301b22c253067197c1c4e208d4e8515