Make query function private and start replacing it with maybe_traced_queries
Summary:
## Context
One of my next goals in the XDB reliability project is T223577767, which is about introducing telemetry into our data layer so we can get the insights we need to shard the DBs and also implement rate limiting.

The MySQL Rust APIs return the stats we need to log and I already started prototyping the changes I'll need to make to pipe that to mononoke_queries and log it. **However**, in this process I noticed how `mononoke_queries` could use a lot of clean up... There's **a lot of duplication/repetition** and the module is a lot more complicated than I think it needs to be...

## This stack
This stack will refactor `mononoke_queries` APIs, to make it easier to understand and maintain.
**More specifically: it will reduce the number of functions defined in each macro block**.

**Current state**:
There are multiple macro match blocks, but since a few make recursive calls, there are only 4 with different implementations of each function. These are:

1. Full read query without cache
2. Full read query with cache
3. Full write query with a list of values
4. Full write query without a list of values

**Each of these blocks** implements 5 functions:
1. traced_query_with_transaction
2. maybe_traced_query_with_transaction
3. query
4. traced_query
5. maybe_traced_query

**My proposal:**
I'll leave these 4 match blocks for now, but I'll reduce the number of functions to 2:
1. query
2. query_with_transaction

Both of these will take an optional `ClientRequestInfo` that the callsites can pass if they want tracing.

I'll do that by basically replacing all the callsites to use the `maybe_traced_*` variants and then removing the `maybe_traced_` prefix after everything has been migrated.

**Notes:**
- I might make some changes throughout the diffs that are redundant when considering the end state. But I decided to leave them instead of redoing the entire stack because the work would be significant, so I didn't deem it worthy considering that each diff will be doing one thing and will be easy to review.
- Between D76196345 and D76196446 I diverge a bit and make changes between `common/rust/shed/sql` and `mononoke_queries` (instead of public APIs like most of the stack). They are of similar nature to what I mentioned above and re-organizing the stack was a lot of work (I tried), so I left them there since they are pretty simple and grouped together 😅

## This diff
- Make the `query` function private and rename it to `query_impl`
- Update all of its callsites to call `maybe_traced_query` (and transaction equivalent).

Reviewed By: markbt

Differential Revision: D76195561

fbshipit-source-id: 15a1228bbf25f7f53fee879e9f3a16ee23c1c7e0