Delete traced_query_with_transaction
Summary:
## Context on this stack
Please see summary of D76195561.

## This diff
Remove the internal `traced_query_with_transaction` function in `mononoke_queries`.
If you see its implementation and how it's used, there's no need for it to exist.
It just calls `query_with_transaction_impl` passing the `cri` if there is one. The other path of the match statement does the same, but passes `None` as the `cri`. So let's just map over the `cri` to pass it if it was provided.

Reviewed By: andreacampi

Differential Revision: D76198581

fbshipit-source-id: 5de09efd37552cac85eb8467c8e4007e6e9d25e7