implement fetch_manifold()
Summary:
This diff adds an implementation of the EdenConfigManager::fetch_manifold method.

I think in the fullness of time, we'll want to create a "FetchConfig" trait with separate concrete implementations for Configerator, Manifold, Interngraph, etc. -- each in a separate file.  If nothing else, this should reduce incremental compile time, since everything won't be in one large file.

At the moment, this only supports Linux, as buck reports something like:
```
This target was marked as incompatible by Buck as part of your changes
```

 because:
```
fbcode//smc/zeus:Allocator is incompatible with cfg:dev-windows-x86_64-msvc-clang19-no-san#9fe5bc7dbf4f8e4a (ovr_config//os:linux unsatisfied), check the target's compatibility attributes
```

We'll have to follow up when it's time to support MacOS and Windows.  In the worst case / short term, we should be able to spawn the manifold CLI to fetch the blob.

Reviewed By: genevievehelsel

Differential Revision: D76367953

fbshipit-source-id: 13e6ea2f9f4e1f88e410d2c64e79a9c79cf9833c