hgevents: add timeout invoking "watchman get-sockname"
Summary:
Apply the client's timeout when trying to get the socket name from "watchman get-sockname".

We typically only fall back to this path when watchman is unexectedly not running (normally watchman is already running and listening on the well known socket). So, there is already an increased chance watchman is unhappy when we hit this code. Under some circumstances I don't fully understand, the "watchman get-sockname" call can block indefinitely. Sapling definitely doesn't want to wait a long time here - we only use watchman for the "hgevents" extension, which is best effort - certainly not worth waiting 10 minutes for.

Differential Revision: D76372784

fbshipit-source-id: 1bc5c0c593bd152cf5491b8ae3fd67581dd7f688