Eden mount reads conflict type and adds events into journal for files changed during force checkout
Summary:
# Bug Description
The eden journal is used to record file change events.
When doing a checkout in force mode, files that are not committed may change if they conflict with files in the new tree. Currently, these files are recorded as an unclean path, but the type of change is not recorded.

Repro:
https://www.internalfb.com/tasks/?t=225707535

# This Diff

Record conflict based file change events into the journal.
Ignore conflicts that don't result in a file change

# Technical Details

# Discussion Points

Reviewed By: genevievehelsel

Differential Revision: D75817781

fbshipit-source-id: 548a6a48136bc7bf98f74d4851bf59dfb50966d3