add dtype to recorded conflicts
Summary:
# Bug Description
The eden journal is used to record file change events.
When doing a checkout in force mode, files that are not committed may change if they conflict with files in the new tree. Currently, these files are recorded as an unclean path, but the type of change is not recorded.

Repro:
https://www.internalfb.com/tasks/?t=225707535

# This Diff

Plumb dtype through addConflict. This will get read in the checkout in EdenMount

# Technical Details

# Discussion Points

Reviewed By: genevievehelsel

Differential Revision: D75817768

fbshipit-source-id: ed5b059f2f16929f6063064db4e4227f3e7362aa