upgrade bytesize to 2.0
Summary:
1.x is getting long in the tooth, and the `to_string` function is smoking some shit.

Behold how KiB and KB are correct, but their natural logs are swapped:

https://www.internalfb.com/code/fbsource/[0833dd526c8ee2423c8614208b69b89c373acad4]/third-party/rust/vendor/bytesize-1.3.0/src/lib.rs?lines=45%2C56%2C68-69

Also behold how `si_prefix = true` gives you IEC, and `si_prefix = false` gives you SI:

https://www.internalfb.com/code/fbsource/[0833dd526c8ee2423c8614208b69b89c373acad4]/third-party/rust/vendor/bytesize-1.3.0/src/lib.rs?lines=182-184%2C190

… or does it?

Look again, and you will see that we will end up taking `1000` to the power of `ln 1024`, and `1024` to the power of `ln 1000` depending on which option you pick. So you're actually getting some complete mathematical nonsense in either case.

https://www.internalfb.com/code/fbsource/[0833dd526c8ee2423c8614208b69b89c373acad4]/third-party/rust/vendor/bytesize-1.3.0/src/lib.rs?lines=45%2C56%2C68-69%2C183-184%2C195-199%2C201-206

I, personally, do not subscribe to the lsd/meth/bath-salts/krokodil version of unit maths, and am electing to codemod this to maintain the 1000 vs 1024 part of this. When in doubt, people get base-2. `dotsync2` has a test that expects SI, so give it that.

Reviewed By: asp2insp

Differential Revision: D76360029

fbshipit-source-id: 3743d8050eb9077c774169d9d48ac7cd01c114f1