subtree: limit the search scope
Summary:
Right now Sapling scans the histories of both `--from-path` and `--to-path` to pick a merge‑base,
which is slow. We’re building a server‑side index to make that fast. Until that ships, workflows
that always merge or copy in one direction can speed things up by searching only one of the
paths’ history.

Reviewed By: muirdm

Differential Revision: D76265844

fbshipit-source-id: 6241b0cb508558057d3e2e05e25acca293878028