Implement `ObjectContent`, an alternative to `OwnedObjectContent` using `ObjectRef`
Summary:
We want to replace the source of truth for Git data with the raw bytes.
This means replacing the "parsed" type from `ObjectRef` to `Object`.
This has become necessary for round-tripping since the latest `gitoxide` upgrade.

We need to associate the lifetime of the `ObjectRef` with that of the backing bytes.
We use `ouroboros` to create a `self_referencing` struct, which allows us to tie the `ObjectRef` lifetime to that of the `ObjectContent` `struct` which holds the bytes.
This restricts how we can access the parsed version. For instance, we can't borrow it because it's not covariant, so we provide enough accessors to do anything we may need to do with it.

We wrap everything in an `Arc` so `ObjectContent` can be `Clone`.
If we can, we'll try to remove this requirement later.

Reviewed By: RajivTS

Differential Revision: D76259407

fbshipit-source-id: 1ae894092e782bf8e3f7e6c1228a284c9ee31d10