Use ObjectRef instead of Object
Summary:
We want to replace the source of truth for Git data with the raw bytes.
This means replacing the "parsed" type from `ObjectRef` to `Object`.
This has become necessary for round-tripping since the latest `gitoxide` upgrade.

This is because the `Object` now represents some kind of canonical representation of Git data which doesn't tolerate "invalid" data, whereas `ObjectRef` is now a thin wrapper around bytes loaded from Git where we are as lenient as Git and guarantee round-tripping by avoiding to parse anything that could be free-form in Git.

So any data that we get from Git (at gitimport time, git push time etc. and that we write to and load from the blobstore) must be represented with the `Ref` types to ensure round-tripping.

Start with replacing `OwnedObjectContent` with `ObjectContent`, which applies this change at the most core area.

This is a minimal change that compiles and passes tests. Follow-ups may be needed.

Reviewed By: RajivTS

Differential Revision: D76259405

fbshipit-source-id: cfa1f70da20a799e7f796011975773e4fc6e994d