Use config backed concurrency parameters for Git Bundle Generator
Summary:
Git Bundle Generator unknowingly made use of the "standard" concurrency parameters for packfile generation. These deviated a while ago from what we set in the config for production services like Mononoke Git server. Let's try to bring the two back in sync. This diff does that.

Once this is landed, we can adjust the concurrency parameters for big bad repos to avoid the possibility of OOMs

I will also land configuration changes which brings the default back to what it was before this diff for non-problematic repos and only keep the reduced values for the problematic ones

Reviewed By: mzr

Differential Revision: D76341546

fbshipit-source-id: 832080798114319b035a9acb8a85050e8e24baff