scs: implement interface for diffing with subtree sources
Summary:
Implement the new `compare_against_subtree_copy_sources` flag in `commit_compare`.

When set, we replace the entries in the parent manifest that correspond to subtree copy destinations with the corresponding entry from the subtree copy source.

This feature can now be used by passing a flag to `scsc`.

Reviewed By: gustavoavena

Differential Revision: D76449506

fbshipit-source-id: b2f3c35f3a462298faeacc2290f6dcb07a4a173d