manifest: implement diffing against subtree copy sources
Summary:
Implement the process of diffing against subtree copies in the manifest diff and ordered_diff implementations.

For subtree copies, we want to replace the original manifest (the "left" side) with the subtree copy source manifest at the point we reach the subtree copy destination.

To allow this, we add a `ReplacementsHolder` struct, which holds a path-tree of replacment manifest entries.  This is traversed in parallel with the manifest we are comparing, and when we hit replacement entries, we replace the left manifest entry with the replacement.

Reviewed By: gustavoavena

Differential Revision: D76449508

fbshipit-source-id: 78b4ce32de1dda7af785f91ce3a3f48a20f934b3