mononoke_api: change interface of ChangesetPathDiffContext
Summary:
`ChangesetPathDiffContext` represents the per-path results of diffing two manifests.  In its current form it can't represent subtree copy comparisons well.  In particular, it can't represent a particular edge-case: a file can be copied (via a subtree copy) but then deleted in the copy.  In a diff against subtree sources, this would show up as a removed file, despite the file not actually existing.

Change the interface of this to prepare for the addition of subtree copy diffs.

Reviewed By: gustavoavena

Differential Revision: D75820141

fbshipit-source-id: 083c6645361860c0fea696d02113b872d213e298