Remove max request size CLI arg for Mononoke Git server
Summary: This argument existed to prevent too large a push from being accepted by Git server. Now this has been moved to a JK in the previous diff so we can delete it. Note that this is safe to delete since this argument was not directly provided a value from TW spec

Differential Revision: D76525335

fbshipit-source-id: bd05ac871feee67d1c96fa50b1c1b2c6bfb49464