move actual config write out of ConfigManager::run()
Summary:
When used as a library (as I'm doing in D76461409), it would be more convienent if ConfigManager::run() returned the serialized config instead of actually writing the config file to disk.

This would allow the caller to control how the config is written.  For example, in Sandcastle's preparer we want to output the config in the log.  We can't do that if everything is wrapped up in an indivisible run() method.

Other potentially useful uses of this is the ability to mutate the serialized config or to use the serialized result in unit tests (a more convienent representation than a TOML string).

This is the first change that substantially changes the structure of the Rust version compared with the Python version.  I think it's probably OK to diverge, especially considering the desire to replace the Python version with the Rust version.

Reviewed By: genevievehelsel

Differential Revision: D76479869

fbshipit-source-id: 4b1304b947d9bef4222a94318fbd0df269d11074