fix a crash during unmount recovery
Summary:
## Context
Accidental unmount recovery can crash eden in some edge cases https://fburl.com/scuba/macos_crash_tailer/mnvo2odz
When we have a failed graceful restart, eden should fall back to the previous eden mount. There are some cases that PriveHelper throw an error and eden mount cannot run successfully. In these cases, the eden mount will unmount and shutdown. If during eden unmount, the accidental unmount recovery runs, it could face to an unexpected situation where the old mount is not in the `EdenServer::mountPoints_` but as it didn't clean up completely yet, it still registered in `EdenMount::mountPoints_`. It is a race condition, and if the unmount recovery runs before the previous unmount finish the `unregisterMount`, eden will crash.
The collected logs from users with this crash shows the failed graceful restart before all of the crashes: P1837166438 or P1837239736

## This diff
Accidental unmount recovery checks the `EdenServer::mountPoints_` and if a path doesn't exist in this map, it consider this as unmount. This diff adds another condition, and we check `EdenMount::mountPoints_` too, and only recover the mount if it is not registered in `EdenMount::mountPoints_`

Reviewed By: MichaelCuevas

Differential Revision: D76475851

fbshipit-source-id: 1a37b49728f54b79e73ef5d44a7df95683f3e47a