remove the CrashPlan advices from doctor
Summary:
Now that CrashPlan crawling issue is solved we can remove the code that specificly check the crawlers and if CrashPlan is in the list it warn users from crawling.

Now that NFS GC works as expected we can run GC and doctor can decrease the number of loaded inodes by running GC. In the next diff, I am going to run GC same as windows. To simplify the code, and use same code for both win and mac, I need to remove this part first

Differential Revision: D76310623

fbshipit-source-id: a48929ee9f7c24e53860db101a1a7e10b85b125a