add NFS GC to doctor when number of inode is high
Summary:
The NFS GC works as expected same as Windows GC, and delete inodes. Now we can add GC to `eden doctor` therefore in addition of periodic GC which runs every hour and with a cutoff of 24 hours delete all inodes which are unused in the last 24 hours, doctor also can run GC if the number of inodes is more than 1 M.
Same as Windows GC, when doctor runs NFS GC the cutoff is 1 hour, and it deletes all inodes if they are not accessed in the last hour. It makes GC doctor more aggressive than  periodic GC.

Differential Revision: D76311748

fbshipit-source-id: 5c3863b28a0f1f99a3381b3132ede5a6be8fa239