rust_cas_client: add boilerplate for bare rust thrift client
Summary:
Context: we're replacing the thin cas client with a bare rust thrift client. This is to get rid of the start up overhead from creating a service router client and to improve reliability by getting rid of the C++ dependencies that the external CAS client pull in.

Add boilerplate for the bare rust thrift client. Name is tentative it will replace the thin client.

Reviewed By: muirdm

Differential Revision: D76593702

fbshipit-source-id: f2f3b025ca22c199bff13283cfcdb4e69c590470