rust_cas_client: implement upload method
Summary:
Context: we're replacing the thin cas client with a bare rust thrift client. This is to get rid of the start up overhead from creating a service router client and to improve reliability by getting rid of the C++ dependencies that the external CAS client pull in.

Implement upload method for the bare rust thrift client.

Reviewed By: liubov-dmitrieva

Differential Revision: D76593863

fbshipit-source-id: 002ed3cc44131654dddf9ecc0c505099e1e4fe11