rust_cas_client: implement fetch method
Summary:
Context: we're replacing the thin cas client with a bare rust thrift client. This is to get rid of the start up overhead from creating a service router client and to improve reliability by getting rid of the C++ dependencies that the external CAS client pull in.

Implement fetch method for the bare rust thrift client.

Reviewed By: liubov-dmitrieva

Differential Revision: D76593860

fbshipit-source-id: 2f447ecd09517af5927f48dc7a23951470d0922a