rust_cas_client: implement prefetch method
Summary:
Context: we're replacing the thin cas client with a bare rust thrift client. This is to get rid of the start up overhead from creating a service router client and to improve reliability by getting rid of the C++ dependencies that the external CAS client pull in.

Implement prefetch method for the bare rust thrift client.

Reviewed By: liubov-dmitrieva

Differential Revision: D76593864

fbshipit-source-id: 78f84ef0f03f1fffc550c9f474802ad9db20e316