Switch to `ObjectContent` for backing packfile items
Summary:
For round-tripping, we need to replace all instances of owned gitoxide types to `ObjectRef`.
When ownership is actually needed, `ObjectContent` is the correct type to use as it stores the owned bytes in addition to the semantic `ObjectRef`.
This is the case for `BaseObject` and `PackfileItem`.

Since we are adding a dependency on `git_types::ObjectContent` for `PackfileItem`, we needed to move the packfile types to become part of `git_types` to avoid creating a circular dependency.

Also expand the `ObjectContent` interface to expose some of the underlying convenience methods on `ObjectRef` to make the call-sites less verbose and require less change.

Reviewed By: RajivTS

Differential Revision: D76604241

fbshipit-source-id: 9c42fa7baddd699e0b205c280ba45be44528bdc1